V3.0.0

See merge request kgcloud/kg-cloud-kgms!308

+++ kgsdk/src/main/java/com/plantdata/kgcloud/plantdata/req/explore/RuleGeneralGraphParameter.java
@@ -11,12 +11,13 @@ import java.util.Map;
 
 /**
  * 普通图探索类
+ *
  * @author Administrator
  */
 @Data
 public class RuleGeneralGraphParameter extends TimeGeneralGraphParameter implements RuleGraphParameter {
 
     private List<RelationReasoningBean> rule;
-    private Map<Integer, JSONObject> reasoningRuleConfigs;
+    private Map<Long, JSONObject> reasoningRuleConfigs;
 
 }
